refactor: sln & csproj identical name + test
