+++ FIPost.sln
@@ -11,8 +11,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LocatieService", "locatie-s
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PakketService", "pakketservice\PakketService.csproj", "{5FBA344B-D270-4637-8CF7-7E77B8A5411E}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "personeel-service", "personeel-service\personeel-service.csproj", "{29F4F469-8EF8-469C-9ACF-B5771333AAC2}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EmployeeService", "EmployeeService\EmployeeService.csproj", "{F8756470-FA08-4AF3-97A2-99F5179319C4}"
 EndProject
 Global
@@ -37,10 +35,6 @@ Global
 		{5FBA344B-D270-4637-8CF7-7E77B8A5411E}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5FBA344B-D270-4637-8CF7-7E77B8A5411E}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{5FBA344B-D270-4637-8CF7-7E77B8A5411E}.Release|Any CPU.Build.0 = Release|Any CPU
-		{29F4F469-8EF8-469C-9ACF-B5771333AAC2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{29F4F469-8EF8-469C-9ACF-B5771333AAC2}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{29F4F469-8EF8-469C-9ACF-B5771333AAC2}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{29F4F469-8EF8-469C-9ACF-B5771333AAC2}.Release|Any CPU.Build.0 = Release|Any CPU
 		{F8756470-FA08-4AF3-97A2-99F5179319C4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{F8756470-FA08-4AF3-97A2-99F5179319C4}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{F8756470-FA08-4AF3-97A2-99F5179319C4}.Release|Any CPU.ActiveCfg = Release|Any CPU
